Add Typer-based create_db and dump_schema modules
