+++ pyproject.toml
@@ -35,6 +35,8 @@ wb_spp_fetch_final = "finmodel.scripts.wb_spp_fetch_final:main"
 wb_spp_fetch_simple = "finmodel.scripts.wb_spp_fetch_simple:main"
 wb_tariffs_box_import = "finmodel.scripts.wb_tariffs_box_import:main"
 wbtariffs_commission_import = "finmodel.scripts.wbtariffs_commission_import:main"
+create_db = "finmodel.scripts.create_db:main"
+dump_schema = "finmodel.scripts.dump_schema:main"
 
 [tool.black]
 line-length = 100
